Re-introduce sorl and bring back ImageBlocks
The 2.0.0 release I made in 2017 had some problems, and wasn't really
the right way to go about things. Let's move forward with the 1.x branch
of pagetree by re-introducing the ImageBlocks. We can work with sorl for
now as we have forked our own version of this library to maintain it for
new Django versions.